Enable MySQL installation and configuration in one run of puppet
"commands :mysql" and "commands :mysqladmin" in the providers
create a restriction: the provider can only be used if those
commands exist. This is reasonable. What is unreasonable is that
this restriction is checked *at puppet startup*, hence the later
availability of MySQL doesn't take effect until next run.

This patch removes those confines - which unfortunately has the
side-effect of requiring explicit calls to execute().

Further, additional dependencies were added to the mysql types
so that they would be satisfied only once the service is installed
and a mysql root user has been created.

A workaround for fixing the situation where a user was deleted
but not dropped (user will not show up in any tables, but you cannot create
them) was added as well.